refactor(mcp): serve the landing page at /ai-plugin instead of /mcp
The page is named /ai-plugin by product decision (Slack, 2026-09-04).
This is explicitly not its permanent home — how the MCP app is
communicated alongside Oak's other open curriculum services is a
broader discussion still to be had.

Moves the route directory and repoints its imports and the deployment
test URL. Component names, the page metadata title and the
/images/mcp/ asset paths are deliberately unchanged: the decision was
about the URL, and the images are already slated to move to Cloudinary
under MCP-684.

Three things at /mcp are deliberately untouched, because they are not
this page:

- public/mcp/carousel/carousel_image_{1,2,3}.png are an external
  contract held by Anthropic's directory listing, guarded by
  mcp-carousel-assets.test.ts and a Playwright spec that asserts NO
  redirect. Any future /mcp redirect must match the exact path only,
  never a wildcard, or it breaks the submission listing.
- www.thenational.academy/mcp is also the MCP endpoint itself, served
  by content negotiation at the Cloudflare edge.
- The live privacy policy links to www.thenational.academy/mcp ten
  times; those links now need either a CMS update or an exact-path
  redirect, and that is tracked separately.
