+++ src/app/(core)/ai-plugin/Components/McpExternalLink.tsx
