+++ src/app/(core)/ai-plugin/Components/McpSection.tsx
