+++ src/app/(core)/ai-plugin/Components/McpTryButton.tsx
