+++ src/app/(core)/ai-plugin/Components/McpView.stories.tsx
