+++ src/app/(core)/ai-plugin/Components/McpView.tsx
@@ -24,7 +24,7 @@ import {
   mcpOutputWarning,
   mcpResponsibleUse,
   mcpSupport,
-} from "@/app/(core)/mcp/mcpContent";
+} from "@/app/(core)/ai-plugin/mcpContent";
 
 /**
  * A titled block of body copy, e.g. "Check the source" followed by a
