+++ src/app/(core)/ai-plugin/mcpContent.ts
