+++ src/app/(core)/ai-plugin/page.tsx
